--- src/main/java/com/maathru/backend/Domain/entity/VisitSchedule.java
@@ -1,37 +1,0 @@
-package com.maathru.backend.Domain.entity;
-
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-import java.time.LocalDate;
-import java.time.LocalDateTime;
-import java.time.LocalTime;
-
-@Entity
-@Table(name = "visit_schedule")
-@AllArgsConstructor
-@NoArgsConstructor
-@Getter
-@Setter
-public class VisitSchedule {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long visitId;
-    private LocalTime time;
-    private LocalDate date;
-    private String type;
-
-    @Column(columnDefinition = "TIMESTAMP DEFAULT CURRENT_TIMESTAMP", insertable = false, updatable = false)
-    private LocalDateTime createdTime;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "midwife_id")
-    private Employee midwife;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "parent_id")
-    private Parent parent;
-}
