+++ src/main/java/com/maathru/backend/Domain/service/RegionService.java
@@ -18,6 +18,7 @@ import com.maathru.backend.enumeration.Province;
 import com.maathru.backend.enumeration.Role;
 import jakarta.transaction.Transactional;
 import lombok.AllArgsConstructor;
+import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.modelmapper.ModelMapper;
 import org.springframework.http.HttpStatus;
@@ -28,7 +29,7 @@ import java.util.List;
 import java.util.stream.Collectors;
 
 @Service
-@AllArgsConstructor
+@RequiredArgsConstructor
 @Slf4j
 public class RegionService {
     private final JwtService jwtService;
