--- src/main/java/com/maathru/backend/External/repository/VisitScheduleRepository.java
@@ -1,7 +1,0 @@
-package com.maathru.backend.External.repository;
-
-import com.maathru.backend.Domain.entity.VisitSchedule;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface VisitScheduleRepository extends JpaRepository<VisitSchedule, Long> {
-}
