+++ src/main/resources/application.properties
@@ -49,8 +49,8 @@ application.security.jwt.access-token-expiration=${JWT_ACCESS_TOKEN_EXPIRATION}
 application.security.jwt.refresh-token-expiration=${JWT_REFRESH_TOKEN_EXPIRATION}
 
 # Define the file name for logging
-logging.file.name=logs/received-logs.log
-logging.level.org.slf4j=INFO
-logging.pattern.console=%d{yyyy-MM-dd HH:mm:ss} %-5level [%thread] %logger{10} [%file:%line] %msg%n
-logging.pattern.file=%d{yyyy-MM-dd HH:mm:ss} %-5level [%thread] %logger{10} [%file:%line] %msg%n
+#logging.file.name=logs/received-logs.log
+#logging.level.org.slf4j=INFO
+#logging.pattern.console=%d{yyyy-MM-dd HH:mm:ss} %-5level [%thread] %logger{10} [%file:%line] %msg%n
+#logging.pattern.file=%d{yyyy-MM-dd HH:mm:ss} %-5level [%thread] %logger{10} [%file:%line] %msg%n
